3rdParty: make mupen64plus-core use 64bit cop0 registers

+++ Source/3rdParty/mupen64plus-core/src/device/pif/bootrom_hle.c
@@ -58,7 +58,7 @@ void pif_bootrom_hle_execute(struct r4300_core* r4300)
     unsigned int tv_type = get_tv_type();   /* 0:PAL, 1:NTSC, 2:MPAL */
 
     int64_t* r4300_gpregs = r4300_regs(r4300);
-    uint32_t* cp0_regs = r4300_cp0_regs(&r4300->cp0);
+    uint64_t* cp0_regs = r4300_cp0_regs(&r4300->cp0);
 
     /* setup CP0 registers */
     cp0_regs[CP0_STATUS_REG] = 0x34000000;
