+++ Source/3rdParty/mupen64plus-core/src/device/r4300/pure_interp.c
@@ -76,7 +76,7 @@ static void InterpretOpcode(struct r4300_core* r4300);
    } \
    static void name##_IDLE(struct r4300_core* r4300, uint32_t op) \
    { \
-      uint32_t* cp0_regs = r4300_cp0_regs(&r4300->cp0); \
+      uint64_t* cp0_regs = r4300_cp0_regs(&r4300->cp0); \
       int* cp0_cycle_count = r4300_cp0_cycle_count(&r4300->cp0); \
       const int take_jump = (condition); \
       if (cop1 && check_cop1_unusable(r4300)) return; \
